Add conditional placement to poster icon buttons
