+++ src/stories/components/ShowPoster.stories.ts
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import { ShowPoster } from '../../components';
 import { MOVIE_DATA } from '../../__tests__/screens/assets';
 import { withRouter } from 'storybook-addon-react-router-v6';
-import { ProfileActions } from '../../hooks/useProfileActions';
+import { ProfileActions } from '../../types';
 
 const profileActionsMock: ProfileActions = {
     removeFromQueue: async () => {},
